requirements.txt
Django>=5.0
python-dotenv>=1.0.0
PyJWT>=2.8
django-cors-headers>=4.0
Pillow>=10.0
mysqlclient>=2.0
PyMySQL>=1.1.0
django-crontab>=0.7.1
openpyxl>=3.1.0
pandas>=2.0.0
openai>=1.59.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
